[Headphones game]: go to ts
